+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SE_2022_Closet">
         <activity
-            android:name=".Activity.JoinActivity"
+            android:name=".startApp_1.activity.JoinActivity"
             android:exported="false" />
         <activity
-            android:name=".Activity.SplashActivity"
+            android:name=".startApp_1.activity.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,12 +22,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Activity.MainActivity"
+            android:name=".startApp_1.activity.MainActivity"
             android:exported="true">
         </activity>
 
         <activity
-            android:name=".Activity.LoginActiyity"
+            android:name=".startApp_1.activity.LoginActiyity"
             android:exported="false"/>
 
     </application>
