+++ app/src/main/java/com/libienz/se_2022_closet/startApp_1/data/Test.java
@@ -1,0 +1,4 @@
+package com.libienz.se_2022_closet.startApp_1.data;
+
+public class Test {
+}
