+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activity.LoginActiyity">
+        tools:context=".startApp_1.activity.LoginActiyity">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -65,7 +65,7 @@
                 android:layout_marginRight="55dp"
                 android:background="#00ff0000"
                 android:text="비밀번호를 잊으셨나요?"
-                android:textColor="#7f6000"
+                android:textColor="#474747"
                 android:textSize="12dp" />
 
             <Button
@@ -78,7 +78,7 @@
                 android:layout_marginRight="55dp"
                 android:background="#00ff0000"
                 android:text="아이디를 잊으셨나요?"
-                android:textColor="#7f6000"
+                android:textColor="#474747"
                 android:textSize="12dp" />
 
             <Button
