+++ app/src/main/res/layout/activity_splash.xml
@@ -8,7 +8,7 @@
     android:background="@color/black"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".Activity.SplashActivity">
+    tools:context=".startApp_1.activity.SplashActivity">
 
 
     <TextView
